pyproject.toml
[tool.poetry]
name = "beemeteo"
version = "0.1.0"
description = "Python library to download weather data."
authors = ["Beedata Analytics <info@beedataanalytics.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-forecastio = "^1.4.0"
pytz = "^2021.3"
click = "^8.0.3"
python-dateutil = "^2.8.2"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
